optional support moved in a separate header
In this way we don't need a define to enable the optional support

The feature detection is copied from
http://en.cppreference.com/w/cpp/experimental/feature_test
and it should work with both experimental/standard optional
